+++ vcxproj2cmake/MSBuildProject.cs
@@ -69,6 +69,7 @@ class MSBuildProject
         var linkLibraryDependenciesXName = XName.Get("LinkLibraryDependencies", msbuildNamespace);
         var linkXName = XName.Get("Link", msbuildNamespace);
         var manifestXName = XName.Get("Manifest", msbuildNamespace);
+        var masmXName = XName.Get("MASM", msbuildNamespace);
         var projectConfigurationXName = XName.Get("ProjectConfiguration", msbuildNamespace);
         var projectReferenceXName = XName.Get("ProjectReference", msbuildNamespace);
         var projectXName = XName.Get("Project", msbuildNamespace);
@@ -101,6 +102,7 @@ class MSBuildProject
                     group.Elements(clCompileXName)
                     .Concat(group.Elements(resourceCompileXName))
                     .Concat(group.Elements(manifestXName))
+                    .Concat(group.Elements(masmXName))
                     .Concat(group.Elements(qtUicXName))
                     .Concat(group.Elements(qtRccXName)))
                 .Select(element => PathUtils.NormalizePathSeparators(UnescapeMSBuildValue(element.Attribute("Include")!.Value.Trim())))
@@ -442,6 +444,7 @@ class MSBuildProject
         "ClCompile",
         "ClInclude",
         "Manifest",
+        "MASM",
         "QtMoc",
         "QtRcc",
         "QtUic",